Players can join tables now

+++ src/main/java/com/silverest/entities/utils/TableCompanion.java
@@ -3,7 +3,7 @@ package com.silverest.entities.utils;
 import com.silverest.entities.Table;
 
 public class TableCompanion {
-    public static Table createTable(int deckCount, int minBet, int maxBet) {
+    public static Table createTable(int deckCount, double minBet, double maxBet) {
         return new Table.TableBuilder()
                 .deckCount(deckCount)
                 .minBet(minBet)
